+++ _codegen/cog_sources/ChunksmithForge.java
@@ -249,6 +249,7 @@ public final class ChunksmithForge {
                 argument(CommandLiteral.SHAPE, string()).suggests(SuggestionProviders.SHAPES));
         registerArguments(command, literal(CommandLiteral.SILENT));
         registerArguments(command, literal(CommandLiteral.SPAWN));
+        registerArguments(command, literal(CommandLiteral.STATUS));
         registerArguments(command, literal(CommandLiteral.START),
                 argument(CommandLiteral.WORLD, dimension()),
                 argument(CommandLiteral.SHAPE, string()).suggests(SuggestionProviders.SHAPES),