+++ _codegen/cog_sources/ChunksmithNeoForge.java
@@ -211,6 +211,7 @@ public class ChunksmithNeoForge {
                 argument(CommandLiteral.SHAPE, string()).suggests(SuggestionProviders.SHAPES));
         registerArguments(command, literal(CommandLiteral.SILENT));
         registerArguments(command, literal(CommandLiteral.SPAWN));
+        registerArguments(command, literal(CommandLiteral.STATUS));
         registerArguments(command, literal(CommandLiteral.START),
                 argument(CommandLiteral.WORLD, dimension()),
                 argument(CommandLiteral.SHAPE, string()).suggests(SuggestionProviders.SHAPES),